Orion_HQ/requirements.txt
numpy
pyyaml
pyadi-iio
skyfield
google-cloud-bigquery
google-cloud-storage
pandas
pytz
